Second Commit

+++ contracts/SupplyChain.sol
@@ -11,8 +11,8 @@ contract SupplyChain {
 
     struct Item {
         string name;
-        uint sku;
-        uint price;
+        uint256 sku;
+        uint256 price;
         State state;
         address payable seller;
         address payable buyer;
